requirements.txt
# LangChain core packages
langchain>=1.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama

# ChromaDB and vector store
langchain-chroma>=0.1.4
chromadb==0.5.20
chroma-hnswlib==0.7.6

# Ollama and utilities
ollama
python-dotenv
sentence-transformers

# TrueLens evaluation
trulens-eval
trulens-providers-langchain
trulens-apps-langchain
